Move setup to index, delete controller
